requirements.txt
psycopg2-binary
requests
apache-airflow
dbt-postgres
